requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
openpyxl==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0

